Brooks dev0 updates

+++ src/main/java/com/sideprojects/trivialpursuit/controller/InvitationController.java
@@ -5,6 +5,7 @@ import javax.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -35,22 +36,20 @@ public class InvitationController {
 	InvitationDAO invitationDAO;
 
 	@RequestMapping(path="/gameboard/{gameCode}/sendInvitation", method=RequestMethod.POST)
-	public String sendInvitation(@RequestParam String username, @PathVariable String gameCode, final HttpServletRequest req, ModelMap map, RedirectAttributes flash) {
+	public String sendInvitation(@RequestParam String username, 
+								@PathVariable String gameCode, 
+								final HttpServletRequest req, 
+								RedirectAttributes flash) {
 		
-		if (username == null || username.length() == 0) {
-			flash.addAttribute("invalidEntry", true);
-			return "redirect:/gameboard/" + gameCode;
-		}
+	    String userIdToken = (String) SessionUtils.get(req, "userIdToken");
+	    User currentUser = userDAO.getUserByToken(userIdToken);
+	    if (currentUser == null) return "redirect:/";
 		
-		// This isn't doing anything as of now  - Brooks
 		Boolean isValid = userDAO.validateUsername(username);
-		if (!isValid) {
-			flash.addAttribute("userNotFound", true);
+		if (!isValid || username == null || username.length() == 0 || username.equals("")) {
+			flash.addFlashAttribute("invalidUser", true);
 			return "redirect:/gameboard/" + gameCode;
 		}
-		
-	    String userId = (String) SessionUtils.get(req, "userIdToken");
-	    User currentUser = userDAO.getUserByToken(userId);
 	    	
 	    invitationDAO.sendInvitation(gameCode, username, currentUser.getUsername());
 		
@@ -60,8 +59,9 @@ public class InvitationController {
 	@RequestMapping(path="/joinGame", method=RequestMethod.POST)
 	public String joinGame(@RequestParam String gameCode, final HttpServletRequest req) {
 		
-	    String userId = (String) SessionUtils.get(req, "userIdToken");
-	    User currentUser = userDAO.getUserByToken(userId);
+	    String userIdToken = (String) SessionUtils.get(req, "userIdToken");
+	    User currentUser = userDAO.getUserByToken(userIdToken);
+	    if (currentUser == null) return "redirect:/";
 	    
     	if (invitationDAO.doesInvitationExist(gameCode, currentUser.getUsername())) {
     		invitationDAO.deleteInvitation(gameCode, currentUser.getUsername());
@@ -100,8 +100,9 @@ public class InvitationController {
 	@RequestMapping(path="/rejectGame", method=RequestMethod.POST)
 	public String rejectGame(@RequestParam String gameCode, final HttpServletRequest req) {
 		
-	    String userId = (String) SessionUtils.get(req, "userIdToken");
-	    User currentUser = userDAO.getUserByToken(userId);
+	    String userIdToken = (String) SessionUtils.get(req, "userIdToken");
+	    User currentUser = userDAO.getUserByToken(userIdToken);
+	    if (currentUser == null) return "redirect:/";
 	    String username = currentUser.getUsername();
 	    
 	    invitationDAO.deleteInvitation(gameCode, username);
@@ -132,8 +133,9 @@ public class InvitationController {
 	@RequestMapping(path="/addFriend", method=RequestMethod.POST)
 	public String addFriend(@RequestParam String username, final HttpServletRequest req) {
 		
-		String userIdToken = (String) SessionUtils.get(req, "userIdToken");
-		User currentUser = userDAO.getUserByToken(userIdToken);
+	    String userIdToken = (String) SessionUtils.get(req, "userIdToken");
+	    User currentUser = userDAO.getUserByToken(userIdToken);
+	    if (currentUser == null) return "redirect:/";
 		
 		invitationDAO.addFriend(currentUser.getUsername(), username);
 		
@@ -143,8 +145,9 @@ public class InvitationController {
 	@RequestMapping(path="/acceptFriendRequest", method=RequestMethod.POST)
 	public String acceptFriendRequest(@RequestParam String username, final HttpServletRequest req) {
 		
-		String userIdToken = (String) SessionUtils.get(req, "userIdToken");
-		User currentUser = userDAO.getUserByToken(userIdToken);
+	    String userIdToken = (String) SessionUtils.get(req, "userIdToken");
+	    User currentUser = userDAO.getUserByToken(userIdToken);
+	    if (currentUser == null) return "redirect:/";
 		
 		invitationDAO.acceptFriendRequest(username, currentUser.getUsername());
 		
@@ -154,8 +157,9 @@ public class InvitationController {
 	@RequestMapping(path="/rejectFriendRequest", method=RequestMethod.POST)
 	public String rejectFriendRequest(@RequestParam String username, final HttpServletRequest req) {
 		
-		String userIdToken = (String) SessionUtils.get(req, "userIdToken");
-		User currentUser = userDAO.getUserByToken(userIdToken);
+	    String userIdToken = (String) SessionUtils.get(req, "userIdToken");
+	    User currentUser = userDAO.getUserByToken(userIdToken);
+	    if (currentUser == null) return "redirect:/";
 		
 		invitationDAO.rejectFriendRequest(currentUser.getUsername(), username);
 		
@@ -165,8 +169,9 @@ public class InvitationController {
 	@RequestMapping(path="/removeFriend", method=RequestMethod.POST)
 	public String removeFriend(@RequestParam String username, final HttpServletRequest req) {
 		
-		String userIdToken = (String) SessionUtils.get(req, "userIdToken");
-		User currentUser = userDAO.getUserByToken(userIdToken);
+	    String userIdToken = (String) SessionUtils.get(req, "userIdToken");
+	    User currentUser = userDAO.getUserByToken(userIdToken);
+	    if (currentUser == null) return "redirect:/";
 		
 		invitationDAO.removeFriend(username, currentUser.getUsername());
 		
@@ -176,8 +181,9 @@ public class InvitationController {
 	@RequestMapping(path="/cancelFriendRequest", method=RequestMethod.POST)
 	public String cancelFriendRequest(@RequestParam String username, final HttpServletRequest req) {
 		
-		String userIdToken = (String) SessionUtils.get(req, "userIdToken");
-		User currentUser = userDAO.getUserByToken(userIdToken);
+	    String userIdToken = (String) SessionUtils.get(req, "userIdToken");
+	    User currentUser = userDAO.getUserByToken(userIdToken);
+	    if (currentUser == null) return "redirect:/";
 		
 		invitationDAO.cancelFriendRequest(currentUser.getUsername(), username);
